requirements.txt
playwright
python-dotenv
beautifulsoup4
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pystray
Pillow
rich
customtkinter
